requirements.txt
pillow
ttkbootstrap
windnd
opencv-python
numpy
